LayoutTests/accessibility/radio-button-group-members-expected.txt


















This tests that linked UI elements withing a radio group work

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Radio buttons within the same named group should be able to access their siblings
PASS test1.linkedUIElementAtIndex(0).isEqual(test1) is true
PASS test1.linkedUIElementAtIndex(1).isEqual(accessibilityController.accessibleElementById('test1_No')) is true
PASS test1.linkedUIElementAtIndex(2).isEqual(accessibilityController.accessibleElementById('test1_Maybe')) is true


Radio buttons that are not in a form, but in the same named group should be able to access their siblings
PASS test2.linkedUIElementAtIndex(0).isEqual(test2) is true
PASS test2.linkedUIElementAtIndex(1).isEqual(accessibilityController.accessibleElementById('test2_No')) is true


Radio buttons labeled by  should be able to access their siblings
PASS test3.linkedUIElementAtIndex(0).isEqual(test3) is true
PASS test3.linkedUIElementAtIndex(1).isEqual(accessibilityController.accessibleElementById('test3_No')) is true


Radio buttons not in a groups should not have siblings (the first linked element is itself)
PASS test4.linkedUIElementAtIndex(0).isEqual(test4) is true
PASS !test4.linkedUIElementAtIndex(1) || !test4.linkedUIElementAtIndex(1).isValid is true


Radio buttons that just have aria roles should work.
PASS ariaRadio1.linkedUIElementAtIndex(0).isEqual(ariaRadio1) is true
PASS ariaRadio1.linkedUIElementAtIndex(1).isEqual(ariaRadio2) is true
PASS successfullyParsed is true

TEST COMPLETE
